Create separate Metronome.js file
